Feat: secure admin views for superuser access only

+++ products/views.py
@@ -3,6 +3,7 @@ from django.urls import reverse
 from django.db.models import Q
 from .models import Category, Product
 from .forms import ProductForm
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 
@@ -61,8 +62,13 @@ def product_details(request, product_id):
     return render(request, "products/product_detail.html", context)
 
 
+@login_required
 def add_product(request):
     """ Add a product to the store """
+
+    if not request.user.is_superuser:
+        return redirect(reverse('products'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -79,8 +85,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+
+    if not request.user.is_superuser:
+        return redirect(reverse('products'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -99,8 +110,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     '''Delete a product from the store'''
+
+    if not request.user.is_superuser:
+        return redirect(reverse('products'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     return redirect(reverse('products'))
